Add call that will initiate a saga in bookinfo.
I found that, contrary to expectations, these services are not actually
backed by databases at all. This is fine - I switched to using in-memory
data for the proof-of-concept.

The `products` service now supports a POST method to `/api/v1/products`
to add new products. This makes a request to `localhost:20000` Our
`qbox` engine on `localhost:3001` should intercept this request,
and make it easy to contact the `details` and `ratings` services to
update details and ratings accordingly.

I have not tested any of this, nor built the corresponding images. Once
this is approved, I will merge, build images, and update the
corresponding Kubernetes YAML to reference these new images.